Got a working soft-exit mechanism
This implies saving the a RestartSimulation at each step, to
deal gracefully with halting the socket loop in the midst of a force
calculation. May want to set a timeout interval to give it a chance
to do one more step.

Also fixed a silly bug where ethermo wasn't restarted