Temporary bug fix for unit stuff
The same problem that started this whole input revamp actually hadn't
been fixed, since cell was being store()'d in fetch(), we undid the
unit conversion, so it was staying in the original units.

I've fixed it by getting rid of the back conversion again, but we might want to
figure out a way to print the units properly.