store now does not do unit change by default.
Previously, we added the functionality that the user could
specify the units to output to the restart file in the code.
However, this means that, if store() is called in fetch(), as the
units are still known at this point store() undoes the unit conversion
of fetch(). Now, unless the units to change into are specified, the
unit conversion is not done, so it should be safer.