Adjusted how default values are printed in the help files
I made it so that instead of printing out self._default,
it instead prints out self.value, which gives the correct
behaviour for classes such as InputCell, for which self._default
is Cell(), whereas self.value = numpy.identity(3). The latter is
in effect the default value of the data between the input tags.