More manual stuff, and fixed the test directories
I've basically got a full outline for the whole new
chapter now.

I've also checked that all the input files in
the test directories run, the only thing left to check is that
the RPMD post-processing files work correctly.

Finally, fixed one bug in the initialization, that meant that you
could not rescale the velocities without also initializing from
file, since one of the variables used in the rescaling was only
initialized in the file initialization routines.