I think I've tidied the driver code fully now
Fixed one or two bugs, gave a name to some floating
point constants used in the code, and made it so that
initialization from the wrapper isn't tried if
initialization from the command line is given. Since
at the moment this is all that is implemented, this
seems sensible to me.