Remove feature branch description
I added the description in the feature branch for exceptions so it is easier to see what the branch is for, but maybe it is better to just open a pull request early and do it there. That way, there is no need to track it by hand and remember to remove the description upon merge.