Open all output safely
Use `open_backup` for all output - for write mode, a backup of the file
that would be overwritten is made. Further, do not open everything in
append mode. If we're at the beginning of the simulation, open in write
mode. This should avoid trajectories with repeated frames in almost all
cases.